add instantiation governance proposal
