+++ cosmwasm/submit-proposal.js
@@ -2,7 +2,14 @@
 
 require('dotenv').config();
 
-const { prepareWallet, prepareClient, submitStoreCodeProposal } = require('./utils');
+const {
+    prepareWallet,
+    prepareClient,
+    submitStoreCodeProposal,
+    submitInstantiateProposal,
+    makeInstantiateMsg,
+    governanceAddress,
+} = require('./utils');
 const { saveConfig, loadConfig, printInfo } = require('../evm/utils');
 
 const { Command, Option } = require('commander');
@@ -22,13 +29,55 @@ const storeCode = (client, wallet, config, options) => {
     });
 };
 
+const instantiate = (client, wallet, config, options, chainName) => {
+    const { contractName } = options;
+    const {
+        axelar: {
+            contracts: { [contractName]: contractConfig },
+        },
+        chains: { [chainName]: chainConfig },
+    } = config;
+
+    const initMsg = makeInstantiateMsg(contractName, chainName, config);
+    submitInstantiateProposal(client, wallet, config, options, initMsg).then(({ proposalId, contractAddress }) => {
+        printInfo('Proposal submitted', proposalId);
+
+        contractConfig.instantiateProposalId = proposalId;
+
+        if (contractAddress) {
+            if (chainConfig) {
+                contractConfig[chainConfig.axelarId] = {
+                    ...contractConfig[chainConfig.axelarId],
+                    address: contractAddress,
+                };
+            } else {
+                contractConfig.address = contractAddress;
+            }
+
+            printInfo(
+                `Predicted address for ${chainName === 'none' ? '' : chainName.concat(' ')}${contractName}. Address`,
+                contractAddress,
+            );
+        }
+    });
+};
+
 const main = async (options) => {
-    const { env } = options;
+    const { env, proposalType } = options;
     const config = loadConfig(env);
 
     await prepareWallet(options)
         .then((wallet) => prepareClient(config, wallet))
-        .then(({ wallet, client }) => storeCode(client, wallet, config, options))
+        .then(({ wallet, client }) => {
+            switch (proposalType) {
+                case 'store':
+                    return storeCode(client, wallet, config, options);
+                case 'instantiate':
+                    return instantiate(client, wallet, config, options);
+                default:
+                    throw new Error('Invalid proposal type');
+            }
+        })
         .then(() => saveConfig(config, env));
 };
 
@@ -48,14 +97,23 @@ const programHandler = () => {
     program.addOption(new Option('-a, --artifactPath <artifactPath>', 'artifact path').makeOptionMandatory(true).env('ARTIFACT_PATH'));
     program.addOption(new Option('-c, --contractName <contractName>', 'contract name').makeOptionMandatory(true));
 
+    program.addOption(new Option('-s, --salt <salt>', 'salt for instantiate2. defaults to contract name').env('SALT'));
+    program.addOption(
+        new Option('--admin <address>', 'when instantiating contract, set an admin address. Defaults to governance module account').default(
+            governanceAddress,
+        ),
+    );
+    program.addOption(new Option('--instantiate2', 'use instantiate2 for constant address deployment'));
     program.addOption(new Option('--aarch64', 'aarch64').env('AARCH64').default(false));
     program.addOption(new Option('-y, --yes', 'skip prompt confirmation').env('YES'));
 
     program.addOption(new Option('-t, --title <title>', 'title of proposal').makeOptionMandatory(true));
     program.addOption(new Option('-d, --description <description>', 'description of proposal').makeOptionMandatory(true));
     program.addOption(new Option('--deposit <deposit>', 'the proposal deposit').makeOptionMandatory(true));
-
     program.addOption(new Option('-r, --runAs <runAs>', 'the address that will execute the message').makeOptionMandatory(true));
+    program.addOption(
+        new Option('--proposalType <proposalType>', 'proposal type').choices(['store', 'instantiate']).makeOptionMandatory(true),
+    );
 
     program.addOption(new Option('--source <source>', "a valid HTTPS URI to the contract's source code"));
     program.addOption(
